requirements.txt
astroid==1.6.0
dj-database-url==0.4.2
Django==2.0
django-cors-headers==2.1.0
django-environ==0.4.4
django-filter==1.1.0
djangorestframework==3.7.7
gunicorn==19.7.1
isort==4.2.15
lazy-object-proxy==1.3.1
Markdown==2.6.10
mccabe==0.6.1
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
pylint==1.8.1
pytz==2017.3
six==1.11.0
whitenoise==3.3.1
wrapt==1.10.11